+++ Libplanet.Store/Trie/TrieExtensions.cs
@@ -5,8 +5,7 @@ using Bencodex.Types;
 
 namespace Libplanet.Store.Trie
 {
-    // FIXME: As it's not an interface, it should be renamed to TrieExtensions.
-    internal static class ITrieExtensions
+    public static class TrieExtensions
     {
         public static ITrie Set(this ITrie trie, IEnumerable<KeyValuePair<KeyBytes, IValue?>> pairs)
         {